# openapi/backend.openapi.yaml
openapi: "3.1.0"
info:
  title: SenseFoundry API
  version: "0.1.0"
  description: AI-assisted Chinese dictionary compilation system API
servers:
  - url: http://localhost:8080/api/v1
security:
  - BearerAuth: []
...
paths:
...
  /submissions:
    get:
      tags: [Submissions]
      operationId: listSubmissions
      parameters: [{$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/PageSize"}, {$ref: "#/components/parameters/Cursor"}]
      responses:
        "200": {$ref: "#/components/responses/SubmissionList"}
        default: {$ref: "#/components/responses/Error"}
    post:
      tags: [Submissions]
      operationId: createSubmission
      x-audit: true
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {$ref: "#/components/requestBodies/SubmissionWrite"}
      responses:
        "201": {$ref: "#/components/responses/SubmissionCreated"}
        default: {$ref: "#/components/responses/Error"}
  /submissions/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    get:
      tags: [Submissions]
      operationId: getSubmission
      responses:
        "200": {$ref: "#/components/responses/Submission"}
        default: {$ref: "#/components/responses/Error"}
...
  /headwords:
    get:
      tags: [Headwords]
      operationId: listHeadwords
      parameters: [{$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/PageSize"}, {$ref: "#/components/parameters/Cursor"}]
      responses:
        "200": {$ref: "#/components/responses/HeadwordList"}
        default: {$ref: "#/components/responses/Error"}
    post:
      tags: [Headwords]
      operationId: createHeadword
      x-audit: true
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {$ref: "#/components/requestBodies/HeadwordWrite"}
      responses:
        "201": {$ref: "#/components/responses/HeadwordCreated"}
        default: {$ref: "#/components/responses/Error"}
...
  /reference-senses:
    get:
      tags: [Reference Senses]
      operationId: listReferenceSenses
      parameters: [{$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/PageSize"}, {$ref: "#/components/parameters/Cursor"}]
      responses:
        "200": {$ref: "#/components/responses/ReferenceSenseList"}
        default: {$ref: "#/components/responses/Error"}
    post:
      tags: [Reference Senses]
      operationId: createReferenceSense
      x-audit: true
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {$ref: "#/components/requestBodies/ReferenceSenseWrite"}
      responses:
        "201": {$ref: "#/components/responses/ReferenceSenseCreated"}
        default: {$ref: "#/components/responses/Error"}
  /reference-senses/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    get:
      tags: [Reference Senses]
      operationId: getReferenceSense
      responses:
        "200": {$ref: "#/components/responses/ReferenceSense"}
        default: {$ref: "#/components/responses/Error"}
    patch:
      tags: [Reference Senses]
      operationId: updateReferenceSense
      x-audit: true
      parameters: [{$ref: "#/components/parameters/OptionalIdempotencyKey"}, {$ref: "#/components/parameters/IfMatch"}]
      requestBody: {$ref: "#/components/requestBodies/ReferenceSenseWrite"}
      responses:
        "200": {$ref: "#/components/responses/ReferenceSense"}
        default: {$ref: "#/components/responses/Error"}
...
  /models:
    get:
      tags: [Models]
      operationId: listModels
      parameters: [{$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/PageSize"}, {$ref: "#/components/parameters/Cursor"}]
      responses:
        "200": {$ref: "#/components/responses/ModelList"}
        default: {$ref: "#/components/responses/Error"}
    post:
      tags: [Models]
      operationId: createModel
      x-audit: true
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {$ref: "#/components/requestBodies/ModelWrite"}
      responses:
        "201": {$ref: "#/components/responses/ModelCreated"}
        default: {$ref: "#/components/responses/Error"}
  /models/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    get:
      tags: [Models]
      operationId: getModel
      responses:
        "200": {$ref: "#/components/responses/Model"}
        default: {$ref: "#/components/responses/Error"}
    patch:
      tags: [Models]
      operationId: updateModel
      x-audit: true
      parameters: [{$ref: "#/components/parameters/OptionalIdempotencyKey"}, {$ref: "#/components/parameters/IfMatch"}]
      requestBody: {$ref: "#/components/requestBodies/ModelWrite"}
      responses:
        "200": {$ref: "#/components/responses/Model"}
        default: {$ref: "#/components/responses/Error"}
  /prompt-templates:
    get:
      tags: [Prompt Templates]
      operationId: listPromptTemplates
      parameters: [{$ref: "#/components/parameters/Page"}, {$ref: "#/components/parameters/PageSize"}, {$ref: "#/components/parameters/Cursor"}]
      responses:
        "200": {$ref: "#/components/responses/PromptTemplateList"}
        default: {$ref: "#/components/responses/Error"}
    post:
      tags: [Prompt Templates]
      operationId: createPromptTemplate
      x-audit: true
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody: {$ref: "#/components/requestBodies/PromptTemplateWrite"}
      responses:
        "201": {$ref: "#/components/responses/PromptTemplateCreated"}
        default: {$ref: "#/components/responses/Error"}
  /prompt-templates/{id}:
    parameters: [{$ref: "#/components/parameters/ResourceId"}]
    get:
      tags: [Prompt Templates]
      operationId: getPromptTemplate
      responses:
        "200": {$ref: "#/components/responses/PromptTemplate"}
        default: {$ref: "#/components/responses/Error"}
    patch:
      tags: [Prompt Templates]
      operationId: updatePromptTemplate
      x-audit: true
      parameters: [{$ref: "#/components/parameters/OptionalIdempotencyKey"}, {$ref: "#/components/parameters/IfMatch"}]
      requestBody: {$ref: "#/components/requestBodies/PromptTemplateWrite"}
      responses:
        "200": {$ref: "#/components/responses/PromptTemplate"}
        default: {$ref: "#/components/responses/Error"}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token issued by Keycloak.
  parameters:
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    Cursor:
      name: cursor
      in: query
      schema: {type: string}
    ResourceId:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 255}
    OptionalIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: {type: string, minLength: 1, maxLength: 255}
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema: {type: string}
  headers:
    ETag:
      description: Entity version identifier.
      schema: {type: string}
    XTraceId:
      description: Request trace identifier.
      schema: {type: string}
  schemas:
    Error:
      type: object
      required: [code, message, details, trace_id]
      properties:
        code: {type: string}
        message: {type: string}
        details: {type: object, additionalProperties: true}
        trace_id: {type: string}
    Pagination:
      type: object
      required: [page, page_size, total]
      properties:
        page: {type: integer}
        page_size: {type: integer}
        total: {type: integer}
        cursor: {type: [string, "null"]}
...
    Submission:
      type: object
      required: [id, document_id, status]
      properties:
        id: {type: string, format: uuid}
        submitter_id: {type: string, format: uuid}
        document_id: {type: string, format: uuid}
        submission_type: {type: string}
        status: {type: string, enum: [pending, processing, accepted, rejected]}
        idempotency_key: {type: string}
        created_at: {type: string, format: date-time}
...
    Headword:
      type: object
      required: [id, character]
      properties:
        id: {type: string, format: uuid}
        character: {type: string}
        pinyin: {type: array, items: {type: string}}
        stroke_count: {type: integer}
        radical: {type: string}
    ReferenceSense:
      type: object
      required: [id, headword_id, gloss]
      properties:
        id: {type: string, format: uuid}
        headword_id: {type: string, format: uuid}
        source_id: {type: string, format: uuid}
        sense_number: {type: integer}
        pos: {type: string}
        gloss: {type: string}
        example_text: {type: string}
        source_kind: {type: string, enum: [authoritative, internal_seed, user_submitted]}
        is_authoritative: {type: boolean, default: false}
        is_publishable: {type: boolean, default: false}
...
    Model:
      type: object
      required: [id, name, version, status]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        version: {type: string}
        model_type: {type: string, enum: [embedding, reranker, classifier, llm]}
        status: {type: string, enum: [registered, active, retired]}
        configuration: {type: object, additionalProperties: true}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    PromptTemplate:
      type: object
      required: [id, name, version, template]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        version: {type: integer}
        template: {type: string}
        variables: {type: array, items: {type: string}}
        active: {type: boolean, default: true}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    SubmissionList: {allOf: [{$ref: "#/components/schemas/PaginatedBase"}, {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/Submission"}}}}]}
...
    HeadwordList: {allOf: [{$ref: "#/components/schemas/PaginatedBase"}, {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/Headword"}}}}]}
    ReferenceSenseList: {allOf: [{$ref: "#/components/schemas/PaginatedBase"}, {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/ReferenceSense"}}}}]}
...
    ModelList: {allOf: [{$ref: "#/components/schemas/PaginatedBase"}, {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/Model"}}}}]}
    PromptTemplateList: {allOf: [{$ref: "#/components/schemas/PaginatedBase"}, {type: object, properties: {data: {type: array, items: {$ref: "#/components/schemas/PromptTemplate"}}}}]}
...
    PaginatedBase:
      type: object
      required: [data, pagination]
      properties:
        data: {type: array, items: {}}
        pagination: {$ref: "#/components/schemas/Pagination"}
  requestBodies:
...
    SubmissionWrite: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/Submission"}}}}
...
    HeadwordWrite: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/Headword"}}}}
    ReferenceSenseWrite: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/ReferenceSense"}}}}
...
    ModelWrite: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/Model"}}}}
    PromptTemplateWrite: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/PromptTemplate"}}}}
  responses:
    Error:
      description: Request failed
      headers: {X-Trace-Id: {$ref: "#/components/headers/XTraceId"}}
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
...
    Submission: {description: Submission, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/Submission"}}}}
    SubmissionCreated: {description: Submission created, headers: {ETag: {$ref: "#/components/headers/ETag"}, Location: {schema: {type: string}}}, content: {application/json: {schema: {$ref: "#/components/schemas/Submission"}}}}
    SubmissionList: {description: Submissions, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/SubmissionList"}}}}
...
    HeadwordCreated: {description: Headword created, headers: {ETag: {$ref: "#/components/headers/ETag"}, Location: {schema: {type: string}}}, content: {application/json: {schema: {$ref: "#/components/schemas/Headword"}}}}
    HeadwordList: {description: Headwords, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/HeadwordList"}}}}
    ReferenceSense: {description: Reference sense, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/ReferenceSense"}}}}
    ReferenceSenseCreated: {description: Reference sense created, headers: {ETag: {$ref: "#/components/headers/ETag"}, Location: {schema: {type: string}}}, content: {application/json: {schema: {$ref: "#/components/schemas/ReferenceSense"}}}}
    ReferenceSenseList: {description: Reference senses, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/ReferenceSenseList"}}}}
...
    Model: {description: Model, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/Model"}}}}
    ModelCreated: {description: Model created, headers: {ETag: {$ref: "#/components/headers/ETag"}, Location: {schema: {type: string}}}, content: {application/json: {schema: {$ref: "#/components/schemas/Model"}}}}
    ModelList: {description: Models, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/ModelList"}}}}
    PromptTemplate: {description: Prompt template, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/PromptTemplate"}}}}
    PromptTemplateCreated: {description: Prompt template created, headers: {ETag: {$ref: "#/components/headers/ETag"}, Location: {schema: {type: string}}}, content: {application/json: {schema: {$ref: "#/components/schemas/PromptTemplate"}}}}
    PromptTemplateList: {description: Prompt templates, headers: {ETag: {$ref: "#/components/headers/ETag"}}, content: {application/json: {schema: {$ref: "#/components/schemas/PromptTemplateList"}}}}
...